Still working on those timers, getting clearer

+++ MiniSumo_2012/delay_ms.c
@@ -2,13 +2,12 @@
  * File Name : delay_ms.c
  * Purpose : For longer delays
  * Creation Date : 26-08-2012
- * Last Modified : sön 26 aug 2012 08:18:18
+ * Last Modified : sön 26 aug 2012 16:11:22
  * Created By : Gabriel Fornaeus, <gf@hax0r.se>
  *
  */
-#include <util/delay.h>
-#include "delay_ms.h"
-void delay_ms(uint16_t ms)
+
+void delay_ms(uint8_t ms)
 {
 	while (ms)
 	{
